vetoproxy-main/backend/requirements.txt
flask
flask-cors
groq
PyPDF2
python-dotenv
requests
gunicorn
